refactor: split drawing to separate file
